requirements.txt
streamlit
streamlit-authenticator
pandas
numpy
tensorflow
scikit-learn
PyYAML
bcrypt
plotly